refactored Users interactors
