+++ app/interactors/users/save.rb
@@ -1,8 +1,20 @@
 module Users
   class Save
-    include Interactor::Organizer
+    include Interactor
 
-    organize Users::Save::Execute,
-             Users::GenerateAccessToken
+    delegate :user_params, to: :context
+
+    def call
+      context.fail! unless user.update(user_params)
+
+      context.user = user
+    end
+
+    private
+
+    def user
+      @user ||= context.user || User.new
+    end
   end
 end
+
